% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_info.R
\name{get_snapshot}
\alias{get_snapshot}
\title{Returns the current snapshotting information}
\usage{
get_snapshot()
}
\value{
Returns the snapshot information as a list
}
\description{
Returns the current snapshotting information
}
\details{
Returns the snapshot information stored in the package. It is a list
with items \code{global}, \code{search}, \code{packages} and \code{namespaces}.
All of these are
character vectors.
}
\seealso{
\code{\link{set_snapshot}}, \code{\link{snapshot}}
}
\author{
Holger Hoefling
}
