% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dep.R
\name{order_pkgs}
\alias{order_pkgs}
\title{Order packages by their dependencies}
\usage{
order_pkgs(pkgs, inst_pkgs, which = c("Depends", "Imports", "LinkingTo"))
}
\arguments{
\item{pkgs}{The packages to order}

\item{inst_pkgs}{a database of installed packages as returned by \code{installed.packages}.}

\item{which}{Which type of dependencies should be taken into account. a character vector listing the types of dependencies, a
subset of ‘c("Depends", "Imports", "LinkingTo", "Suggests",
"Enhances")’.  Character string ‘"all"’ is shorthand for that
vector, character string ‘"most"’ for the same vector without
‘"Enhances"’.}

\item{inst_pkgs}{The installed packages as returned by \code{installed.packages}}
}
\value{
A data frame with the column 'pkg' giving the name of the package and 'level', by which the dataframe is sorted from lowest to highest.
}
\description{
Order packages by their dependencies
}
\details{
Given packages, the function orders them so that every package
only depends on other packages with lower order number. This way,
they can safely be removed from the highest value down.
}
\author{
Holger Hoefling
}
