% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanWorkspace.R
\name{sanitize}
\alias{sanitize}
\alias{cleanWorkspace}
\title{Cleans the workspace}
\usage{
sanitize(keep_global = keep$global, keep_search = keep$search,
  keep_packages = keep$packages, keep_namespaces = keep$namespaces,
  hidden = FALSE, verbose = TRUE)

cleanWorkspace(keep_global = keep$global, keep_search = keep$search,
  keep_packages = keep$packages, keep_namespaces = keep$namespaces,
  hidden = FALSE, verbose = TRUE)
}
\arguments{
\item{keep_global}{Which global variables to keep}

\item{keep_search}{Which environments to keep on the search path}

\item{keep_packages}{Which loaded packages to keep}

\item{keep_namespaces}{Which loaded namespaces to keep}

\item{hidden}{Should also hidden variables starting with a dot be removed?}

\item{verbose}{Should it be verbose if packages can't be undetached?}
}
\value{
TRUE if everything succeeded, FALSE otherwise
}
\description{
Cleans the current workspace
}
\details{
Removes all elements in the global environment and detaches all attached packages that are not base packages.
One can specify certain packages that should be kept.
}
\author{
Holger Hoefling
}
