% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_info.R
\name{set_snapshot}
\alias{set_snapshot}
\title{Sets the saved snapshot}
\usage{
set_snapshot(snap)
}
\arguments{
\item{snap}{An object structured as the return of the
\code{\link{snapshot}} function. It is a list with elements \code{global},
\code{search}, \code{packages} and \code{namespaces}. All of these are
character vectors.}
}
\value{
Invisibly, the input \code{snap}.
}
\description{
Sets the saved snapshot
}
\details{
Sets the saved information in the package to the contents
of the \code{snap} parameter
}
\seealso{
\code{\link{get_snapshot}}, \code{\link{snapshot}}
}
\author{
Holger Hoefling
}
