% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot}
\alias{snapshot}
\title{Snapshot current configuration}
\usage{
snapshot(global = TRUE, search = TRUE, packages = TRUE,
  namespaces = TRUE)
}
\arguments{
\item{global}{Snapshot global variables?}

\item{search}{Snapshot state of search path?}

\item{packages}{Snapshot loaded packages}

\item{namespaces}{Snapshot loaded namespaces?}
}
\value{
list with the stored items; invisibly
}
\description{
Snapshot current configuration
}
\details{
Stores which global variables currently exist, what environments or data.frames are attached to the
search path, which packages are loaded and which namespaces are loaded
}
\seealso{
\code{\link{get_snapshot}}, \code{\link{set_snapshot}}
}
\author{
Holger Hoefling
}
